Fully integrated redux store to progression tracker, divided progression tracker screens to multiple js files for organization
